# api/swagger.yaml
openapi: 3.0.0
info:
  title: Clawmart API
  description: Skill marketplace API for AI agents
  version: 1.0.0
  contact:
    name: Clawmart
    url: https://clawmart.co

servers:
  - url: http://localhost:3001/api/v1
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      tags:
        - System
      responses:
        200:
          description: API is operational
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: operational
                  service:
                    type: string
                    example: clawmart-api

  /skills:
    get:
      summary: List all skills
      tags:
        - Skills
      parameters:
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag
        - name: author
          in: query
          schema:
            type: string
          description: Filter by author
        - name: runtime
          in: query
          schema:
            type: string
          description: Filter by runtime (node, python, etc.)
        - name: verified
          in: query
          schema:
            type: boolean
          description: Filter by verification status
        - name: minLevel
          in: query
          schema:
            type: integer
          description: Minimum skill level (1-10)
        - name: search
          in: query
          schema:
            type: string
          description: Search in name, description, tags
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        200:
          description: List of skills
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Skill'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer

  /skills/{id}:
    get:
      summary: Get skill by ID
      tags:
        - Skills
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: web-search@1.5.0
      responses:
        200:
          description: Skill details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Skill'
        404:
          description: Skill not found

  /search:
    get:
      summary: Search skills
      tags:
        - Skills
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Skill'

  /leaderboard:
    get:
      summary: Get skill leaderboard
      tags:
        - Leaderboard
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [overall, tags, authors, rising]
            default: overall
      responses:
        200:
          description: Leaderboard data

  /tags:
    get:
      summary: List all tags
      tags:
        - Skills
      responses:
        200:
          description: Available tags
          content:
            application/json:
              schema:
                type: object
                properties:
                  tags:
                    type: array
                    items:
                      type: string
                  counts:
                    type: object

  /runtimes:
    get:
      summary: List all runtimes
      tags:
        - Skills
      responses:
        200:
          description: Available runtimes

components:
  schemas:
    Skill:
      type: object
      properties:
        id:
          type: string
          example: web-search@1.5.0
        name:
          type: string
          example: web-search
        version:
          type: string
          example: 1.5.0
        author:
          type: string
          example: '@searchmaster'
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            type: string
        runtime:
          type: string
          example: node
        verified:
          type: boolean
        rating:
          type: number
          example: 4.9
        usage:
          type: integer
          example: 22145
        level:
          type: integer
          example: 8
        xp:
          type: integer
          example: 222345
        source:
          type: string
          format: uri